requirements.txt
numpy==1.23.3
pandas==1.5.1
h5py==3.7.0
seaborn==0.12.1
scikit-learn==1.1.2
tables==3.7.0
scikit-misc==0.1.4
annoy
joblib==1.2.0
scanpy==1.9.1
umap
fbpca
scib==1.1.4
harmony-pytorch==0.1.7
matplotlib==3.6.0